% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\docType{data}
\name{gh}
\alias{gh}
\title{gh}
\format{
A \code{tibble} with 4 columns which are :
\describe{
\item{y1}{Observation on first timepoint}
\item{y2}{Observation on second timepoint}
\item{y3}{Observation on first timepoint}
\item{Treatment}{Treatment arm of the patient}}
}
\usage{
data(gh)
}
\description{
High dimensional data on three consecutive measurements for and treatment arm information column.
}
\keyword{datasets}
