% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitdata.R
\docType{data}
\name{longitdata}
\alias{longitdata}
\title{Repeatedly measured protein expression data}
\format{
A \code{tibble} with 2 columns which are :
\describe{
\item{ID}{Patient ID}
\item{y}{Repeated observations on the patient arranged in a row as per a subject}}
}
\usage{
data(longitdata)
}
\description{
Longitudinal observation on single variable from different observations. Observations arranged in a column as the patient with corresponding column of ID.
}
\keyword{datasets}
